+++ src/Examples/AlexNet.cs
@@ -63,7 +63,7 @@ namespace TorchSharp.Examples
             using (var f = features.forward(input))
             using (var avg = avgPool.forward(f))
 
-            using (var x = avg.view(new long[] { avg.shape[0], 256 * 2 * 2 }))
+            using (var x = avg.reshape(new long[] { avg.shape[0], 256 * 2 * 2 }))
                 return classifier.forward(x);
         }
     }