+++ src/Examples/MNIST.cs
@@ -30,7 +30,7 @@ namespace TorchSharp.Examples
     /// </remarks>
     public class MNIST
     {
-        private static int _epochs = 4;
+        private static int _epochs = 1;
         private static int _trainBatchSize = 64;
         private static int _testBatchSize = 128;
 
@@ -95,6 +95,7 @@ namespace TorchSharp.Examples
                 Test(model, nll_loss(reduction: torch.nn.Reduction.Sum), device, test, test.Size);
 
                 Console.WriteLine($"End-of-epoch memory use: {GC.GetTotalMemory(false)}");
+                scheduler.step();
             }
 
             sw.Stop();
@@ -135,21 +136,21 @@ namespace TorchSharp.Examples
 
             public override Tensor forward(Tensor input)
             {
-                var l11 = conv1.forward(input);
-                var l12 = relu1.forward(l11);
+                using var l11 = conv1.forward(input);
+                using var l12 = relu1.forward(l11);
 
-                var l21 = conv2.forward(l12);
-                var l22 = relu2.forward(l21);
-                var l23 = pool1.forward(l22);
-                var l24 = dropout1.forward(l23);
+                using var l21 = conv2.forward(l12);
+                using var l22 = relu2.forward(l21);
+                using var l23 = pool1.forward(l22);
+                using var l24 = dropout1.forward(l23);
 
-                var x = flatten.forward(l24);
+                using var x = flatten.forward(l24);
 
-                var l31 = fc1.forward(x);
-                var l32 = relu3.forward(l31);
-                var l33 = dropout2.forward(l32);
+                using var l31 = fc1.forward(x);
+                using var l32 = relu3.forward(l31);
+                using var l33 = dropout2.forward(l32);
 
-                var l41 = fc2.forward(l33);
+                using var l41 = fc2.forward(l33);
 
                 return logsm.forward(l41);
             }