+++ src/TorchSharp/NN/Sequential.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Runtime.InteropServices;
+
 using static TorchSharp.torch;
 
 namespace TorchSharp
@@ -36,6 +37,11 @@ namespace TorchSharp
                 _names.Add(name);
             }
 
+            public void Add(torch.nn.Module module)
+            {
+                Add(_modules.Count.ToString(), module);
+            }
+
             internal Sequential(IntPtr handle) : base(handle, IntPtr.Zero)
             {
             }
@@ -45,17 +51,29 @@ namespace TorchSharp
 
             public override Tensor forward(Tensor tensor)
             {
-                if (_modules.Count == 0) return tensor.clone();
+                // If there are no modules, just return a fresh handle to the input.
+                if (_modules.Count == 0) return tensor.alias();
+
+                // The loop-based logic below only works for n > 1, so here's another special case.
+                if (_modules.Count == 1) return _modules[0].forward(tensor);
 
                 // Note: we have not been able to detect any significant performance difference between
                 // implementing forward() in native or managed code.
 
                 // Using an for loop helps debugging, since we can know the ordinal of the submodule.
 
-                for (var idx = 0; idx < _modules.Count; idx++) {
-                    tensor = _modules[idx].forward(tensor);
+                var t0 = _modules[0].forward(tensor);
+
+                for (var idx = 1; idx < _modules.Count-1; idx++) {
+                    var t1 = _modules[idx].forward(t0);
+                    t0.Dispose();
+                    t0 = t1;
                 }
-                return tensor;
+
+                var result = _modules[_modules.Count - 1].forward(t0);
+                t0.Dispose();
+
+                return result;
             }
 
             public override nn.Module apply(Action<nn.Module> fn)
@@ -65,6 +83,12 @@ namespace TorchSharp
                 fn(this);
                 return this;
 
+            }
+
+            protected override void Dispose(bool disposing)
+            {
+                foreach (var m in _modules) { m.Dispose(); }
+                base.Dispose(disposing);
             }
 
             // Currently, Sequential is implemented entirely in managed code, but if we go back to
@@ -86,6 +110,18 @@ namespace TorchSharp
             extern static IntPtr THSNN_Sequential_ctor();
 
             /// <summary>
+            /// Get empty sequential
+            /// </summary>
+            /// <returns></returns>
+            static public Sequential Sequential()
+            {
+                var handle = THSNN_Sequential_ctor();
+                if(handle == IntPtr.Zero) {torch.CheckForErrors();}
+
+                return new Sequential(handle);
+            }
+
+            /// <summary>
             /// A sequential container. Modules will be added to it in the order they are passed in the constructor.
             /// Alternatively, an OrderedDict of modules can be passed in. The forward() method of Sequential accepts any input and forwards it to the first module it contains.
             /// It then “chains” outputs to inputs sequentially for each subsequent module, finally returning the output of the last module.
@@ -94,11 +130,61 @@ namespace TorchSharp
             /// </summary>
             /// <param name="modules">An ordered list of the contained modules.</param>
             /// <returns></returns>
+            /// <remarks>Sequential will take ownership of the modules and dispose of them when disposed.</remarks>
             static public Sequential Sequential(params (string name, torch.nn.Module submodule)[] modules)
             {
-                var handle = THSNN_Sequential_ctor();
-                if (handle == IntPtr.Zero) { torch.CheckForErrors(); }
-                var res = new Sequential(handle);
+                var res = Sequential();
+                foreach (var module in modules)
+                    res.Add(module.name, module.submodule);
+                return res;
+            }
+
+            /// <summary>
+            /// A sequential container. Modules will be added to it in the order they are passed in the constructor.
+            /// It then “chains” outputs to inputs sequentially for each subsequent module, finally returning the output of the last module.
+            /// The value a Sequential provides over manually calling a sequence of modules is that it allows treating the whole container as a single module,
+            /// such that performing a transformation on the Sequential applies to each of the modules it stores (which are each a registered submodule of the Sequential).
+            /// </summary>
+            /// <param name="modules">An ordered list of the contained modules.</param>
+            /// <returns></returns>
+            /// <remarks>Sequential will take ownership of the modules and dispose of them when disposed.</remarks>
+            static public Sequential Sequential(params torch.nn.Module[] modules)
+            {
+                var res = Sequential();
+                foreach(var m in modules)
+                    res.Add(m);
+                return res;
+            }
+
+            /// <summary>
+            /// A sequential container. Modules will be added to it in the order they are passed in the constructor.
+            /// Alternatively, an OrderedDict of modules can be passed in. The forward() method of Sequential accepts any input and forwards it to the first module it contains.
+            /// It then “chains” outputs to inputs sequentially for each subsequent module, finally returning the output of the last module.
+            /// The value a Sequential provides over manually calling a sequence of modules is that it allows treating the whole container as a single module,
+            /// such that performing a transformation on the Sequential applies to each of the modules it stores (which are each a registered submodule of the Sequential).
+            /// </summary>
+            /// <param name="modules">An ordered list of the contained modules.</param>
+            /// <remarks>Sequential will take ownership of the modules and dispose of them when disposed.</remarks>
+            static public Sequential Sequential(params System.Tuple<string, torch.nn.Module>[] modules)
+            {
+                var res = Sequential();
+                foreach (var module in modules)
+                    res.Add(module.Item1, module.Item2);
+                return res;
+            }
+
+            /// <summary>
+            /// A sequential container. Modules will be added to it in the order they are passed in the constructor.
+            /// Alternatively, an OrderedDict of modules can be passed in. The forward() method of Sequential accepts any input and forwards it to the first module it contains.
+            /// It then “chains” outputs to inputs sequentially for each subsequent module, finally returning the output of the last module.
+            /// The value a Sequential provides over manually calling a sequence of modules is that it allows treating the whole container as a single module,
+            /// such that performing a transformation on the Sequential applies to each of the modules it stores (which are each a registered submodule of the Sequential).
+            /// </summary>
+            /// <param name="modules">An ordered list of the contained modules.</param>
+            /// <remarks>Sequential will take ownership of the modules and dispose of them when disposed.</remarks>
+            static public Sequential Sequential(IEnumerable<(string name, torch.nn.Module submodule)> modules)
+            {
+                var res = Sequential();
                 foreach (var module in modules)
                     res.Add(module.name, module.submodule);
                 return res;
@@ -112,17 +198,31 @@ namespace TorchSharp
             /// such that performing a transformation on the Sequential applies to each of the modules it stores (which are each a registered submodule of the Sequential).
             /// </summary>
             /// <param name="modules">An ordered list of the contained modules.</param>
-            static public Sequential Sequential(IEnumerable<(string name, torch.nn.Module submodule)> modules)
+            /// <remarks>Sequential will take ownership of the modules and dispose of them when disposed.</remarks>
+            static public Sequential Sequential(IEnumerable<System.Tuple<string, torch.nn.Module>> modules)
             {
-                var handle = THSNN_Sequential_ctor();
-                if (handle == IntPtr.Zero) { torch.CheckForErrors(); }
-                var res = new Sequential(handle);
+                var res = Sequential();
                 foreach (var module in modules)
-                    res.Add(module.name, module.submodule);
+                    res.Add(module.Item1, module.Item2);
                 return res;
             }
 
+            /// <summary>
+            /// A sequential container. Modules will be added to it in the order they are passed in the constructor.
+            /// It then “chains” outputs to inputs sequentially for each subsequent module, finally returning the output of the last module.
+            /// The value a Sequential provides over manually calling a sequence of modules is that it allows treating the whole container as a single module,
+            /// such that performing a transformation on the Sequential applies to each of the modules it stores (which are each a registered submodule of the Sequential).
+            /// </summary>
+            /// <param name="modules">An ordered list of the contained modules.</param>
+            /// <returns></returns>
+            /// <remarks>Sequential will take ownership of the modules and dispose of them when disposed.</remarks>
+            static public Sequential Sequential(IEnumerable<torch.nn.Module> modules)
+            {
+                var res = Sequential();
+                foreach(var module in modules)
+                    res.Add(module);
+                return res;
+            }
         }
-
     }
 }